+++ src/HCFS/dir_entry_btree.h
@@ -15,28 +15,28 @@
 
 #include "fuseop.h"
 
-int dentry_binary_search(const DIR_ENTRY *entry_array, const int num_entries,
-			const DIR_ENTRY *new_entry, int *index_to_insert);
+int32_t dentry_binary_search(const DIR_ENTRY *entry_array, const int32_t num_entries,
+			const DIR_ENTRY *new_entry, int32_t *index_to_insert);
 
-int search_dir_entry_btree(const char *target_name, DIR_ENTRY_PAGE *tnode,
-		int fh, int *result_index, DIR_ENTRY_PAGE *result_node);
+int32_t search_dir_entry_btree(const char *target_name, DIR_ENTRY_PAGE *tnode,
+		int32_t fh, int32_t *result_index, DIR_ENTRY_PAGE *result_node);
 
 /* if returns 1, then there is an entry to be added to the parent */
-int insert_dir_entry_btree(DIR_ENTRY *new_entry, DIR_ENTRY_PAGE *tnode,
-	int fh, DIR_ENTRY *overflow_median, long long *overflow_new_page,
+int32_t insert_dir_entry_btree(DIR_ENTRY *new_entry, DIR_ENTRY_PAGE *tnode,
+	int32_t fh, DIR_ENTRY *overflow_median, int64_t *overflow_new_page,
 	DIR_META_TYPE *this_meta, DIR_ENTRY *tmp_entries,
-	long long *temp_child_page_pos);
+	int64_t *temp_child_page_pos);
 
-int delete_dir_entry_btree(DIR_ENTRY *to_delete_entry, DIR_ENTRY_PAGE *tnode,
-	int fh, DIR_META_TYPE *this_meta, DIR_ENTRY *tmp_entries,
-	long long *temp_child_page_pos);
+int32_t delete_dir_entry_btree(DIR_ENTRY *to_delete_entry, DIR_ENTRY_PAGE *tnode,
+	int32_t fh, DIR_META_TYPE *this_meta, DIR_ENTRY *tmp_entries,
+	int64_t *temp_child_page_pos);
 
-int rebalance_btree(DIR_ENTRY_PAGE *tnode, int fh, DIR_META_TYPE *this_meta,
-	int selected_child, DIR_ENTRY *tmp_entries,
-	long long *temp_child_page_pos);
+int32_t rebalance_btree(DIR_ENTRY_PAGE *tnode, int32_t fh, DIR_META_TYPE *this_meta,
+	int32_t selected_child, DIR_ENTRY *tmp_entries,
+	int64_t *temp_child_page_pos);
 
-int extract_largest_child(DIR_ENTRY_PAGE *tnode, int fh,
+int32_t extract_largest_child(DIR_ENTRY_PAGE *tnode, int32_t fh,
 	DIR_META_TYPE *this_meta, DIR_ENTRY *extracted_child,
-	DIR_ENTRY *tmp_entries, long long *temp_child_page_pos);
+	DIR_ENTRY *tmp_entries, int64_t *temp_child_page_pos);
 
 #endif  /* GW20_HCFS_DIR_ENTRY_BTREE_H_ */
