+++ src/HCFS/metaops.h
@@ -29,44 +29,44 @@
 #include "filetables.h"
 #include "mount_manager.h"
 
-int dir_add_entry(ino_t parent_inode, ino_t child_inode, const char *childname,
+int32_t dir_add_entry(ino_t parent_inode, ino_t child_inode, const char *childname,
 			mode_t child_mode, META_CACHE_ENTRY_STRUCT *body_ptr);
-int dir_remove_entry(ino_t parent_inode, ino_t child_inode,
+int32_t dir_remove_entry(ino_t parent_inode, ino_t child_inode,
 			const char *childname,
 			mode_t child_mode, META_CACHE_ENTRY_STRUCT *body_ptr);
-int change_parent_inode(ino_t self_inode, ino_t parent_inode1,
+int32_t change_parent_inode(ino_t self_inode, ino_t parent_inode1,
 			ino_t parent_inode2, META_CACHE_ENTRY_STRUCT *body_ptr);
-int decrease_nlink_inode_file(fuse_req_t req, ino_t this_inode);
-int change_dir_entry_inode(ino_t self_inode, const char *targetname,
+int32_t decrease_nlink_inode_file(fuse_req_t req, ino_t this_inode);
+int32_t change_dir_entry_inode(ino_t self_inode, const char *targetname,
 			ino_t new_inode,
 			mode_t new_mode, META_CACHE_ENTRY_STRUCT *body_ptr);
-int delete_inode_meta(ino_t this_inode);
-int init_dir_page(DIR_ENTRY_PAGE *tpage, ino_t self_inode, ino_t parent_inode,
-						long long this_page_pos);
-long long seek_page(META_CACHE_ENTRY_STRUCT *body_ptr, long long target_page,
-			long long hint_page);
-long long create_page(META_CACHE_ENTRY_STRUCT *body_ptr, long long target_page);
+int32_t delete_inode_meta(ino_t this_inode);
+int32_t init_dir_page(DIR_ENTRY_PAGE *tpage, ino_t self_inode, ino_t parent_inode,
+						int64_t this_page_pos);
+int64_t seek_page(META_CACHE_ENTRY_STRUCT *body_ptr, int64_t target_page,
+			int64_t hint_page);
+int64_t create_page(META_CACHE_ENTRY_STRUCT *body_ptr, int64_t target_page);
 
-long long seek_page2(FILE_META_TYPE *temp_meta, FILE *fptr,
-		long long target_page, long long hint_page);
+int64_t seek_page2(FILE_META_TYPE *temp_meta, FILE *fptr,
+		int64_t target_page, int64_t hint_page);
 
-int actual_delete_inode(ino_t this_inode, char d_type, ino_t root_inode,
+int32_t actual_delete_inode(ino_t this_inode, char d_type, ino_t root_inode,
 			MOUNT_T *mptr);
-int mark_inode_delete(fuse_req_t req, ino_t this_inode);
+int32_t mark_inode_delete(fuse_req_t req, ino_t this_inode);
 
-int disk_markdelete(ino_t this_inode, MOUNT_T *mptr);
-int disk_cleardelete(ino_t this_inode, ino_t root_inode);
-int disk_checkdelete(ino_t this_inode, ino_t root_inode);
-int startup_finish_delete(void);
+int32_t disk_markdelete(ino_t this_inode, MOUNT_T *mptr);
+int32_t disk_cleardelete(ino_t this_inode, ino_t root_inode);
+int32_t disk_checkdelete(ino_t this_inode, ino_t root_inode);
+int32_t startup_finish_delete(void);
 
-int lookup_dir(ino_t parent, const char *childname, DIR_ENTRY *dentry);
-int change_pin_flag(ino_t this_inode, mode_t this_mode, char new_pin_status);
+int32_t lookup_dir(ino_t parent, const char *childname, DIR_ENTRY *dentry);
+int32_t change_pin_flag(ino_t this_inode, mode_t this_mode, char new_pin_status);
 
-int collect_dir_children(ino_t this_inode, ino_t **dir_node_list,
-	long long *num_dir_node, ino_t **nondir_node_list,
-	long long *num_nondir_node);
+int32_t collect_dir_children(ino_t this_inode, ino_t **dir_node_list,
+	int64_t *num_dir_node, ino_t **nondir_node_list,
+	int64_t *num_nondir_node);
 
-int inherit_xattr(ino_t parent_inode, ino_t this_inode,
+int32_t inherit_xattr(ino_t parent_inode, ino_t this_inode,
 		META_CACHE_ENTRY_STRUCT *selbody_ptr);
 
 int change_unpin_dirty_size(ino_t this_inode, char ispin);
